move LinkCard to components common folder
